Fixing some warnings

+++ application/src/views/ActiveSessionView.cpp
@@ -57,7 +57,7 @@ ActiveSessionView::ActiveSessionView(
     json& jActiveSessionView = cfg["activesessionview"];
     json& jExpanded = jActiveSessionView["expanded"];
     if (jExpanded.is_array())
-        for (int i = 0; i != jExpanded.size() && i < metadataEditWidgets_.size(); ++i)
+        for (int i = 0; i != (int)jExpanded.size() && i < metadataEditWidgets_.size(); ++i)
             if (jExpanded[i].is_boolean() && jExpanded[i].get<bool>())
                 metadataEditWidgets_[i]->setExpanded(true);
